Move files that aren't task scripts or config to lib/ or tasks/
This almost completely solves #18. Files in `lib/` are at the top of the
abstraction; files in `lib/input` are functions that handle keyboard
events; files in `lib/draw` are functions that handle the display during
task phases; `lib/phases` are functions that handle the task phases
themselves; and `lib/*/helpers` hosts small auxilliary functions for
each file category.

Task scripts and config might be moved later.

The distinction between `lib/draw` and `lib/phase` is the first small
change to address #37.