Re-structure Data.blocks; introduce runNthBlock (#39)
`runNthBlock` is introduced to eliminate duplication -- since per-block
config will be included in DataObject.blocks{k}.config, there's no need
to pass it again. (At least not to the block runner. The block runner
does extract it, but then again, none of the abstraction levels
underneath it interface with the DataObject directly.)

HLFF is a partial proof of concept. It will only become a full proof of
concept once the condition data structure is introduced (#93).