Introduce config option `blocksPerCondition`
In practice, `blockConfig` is "configuration for all blocks within a
particular condition." The `numBlocks` option was not appropriate for
this task, since the existing code looks to it to determine how many
blocks *in total* there will be.

This also introduces basic tests for the trials generated for specific
tasks. (Later, the tests should probably be split up by task and go into
task-specific folders.) Right now, given the symmetry of conditions in
all tasks, the tests look at a single condition only.